feat: split code in different concerns
